requirements.txt
boto3==1.9.86
botocore==1.12.86
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
docopt==0.6.2
docutils==0.14
html5-parser==0.4.5
idna==2.8
jmespath==0.9.3
lxml==4.3.0
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
rstcloth==0.2.6
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
